Move test to general tester
